# docs/openapi/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "TalkRealm 是一個即時通訊平台，提供文字和語音聊天功能。",
        "title": "TalkRealm API",
        "termsOfService": "http://talkrealm.example.com/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.talkrealm.example.com/support",
            "email": "support@talkrealm.example.com"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/learn/campaign/leaderboard": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Learn"
                ],
                "summary": "取得固定關卡排行榜",
                "parameters": [
                    {
                        "type": "string",
                        "description": "friends 表示好友榜",
                        "name": "scope",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/service.LeaderboardView"
                        }
                    }
                }
            }
        },
        "/api/v1/learn/campaign/{no}": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Learn"
                ],
                "summary": "開始固定關卡",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "關卡編號",
                        "name": "no",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "語系",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.startCampaignReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/service.CrosswordView"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/learn/leaderboard/weekly": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Learn"
                ],
                "summary": "取得本週 XP 排行榜",
                "parameters": [
                    {
                        "type": "string",
                        "description": "friends 表示好友榜",
                        "name": "scope",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/service.LeaderboardView"
                        }
                    }
                }
            }
        },
        "/api/v1/learn/levels": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Learn"
                ],
                "summary": "生成單字關卡",
                "parameters": [
                    {
                        "description": "模式與難度",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.createLevelReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/service.LevelView"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/learn/levels/crossword": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Learn"
                ],
                "summary": "生成交叉字謎網格關卡",
                "parameters": [
                    {
                        "description": "難度",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.createCrosswordReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/service.CrosswordView"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/learn/levels/{id}/guess": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Learn"
                ],
                "summary": "提交單字作答",
                "parameters": [
                    {
                        "type": "string",
                        "description": "關卡 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "作答內容",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.LearnGuessRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/service.GuessOutcome"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "410": {
                        "description": "Gone",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/learn/levels/{id}/hint": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Learn"
                ],
                "summary": "單字提示",
                "parameters": [
                    {
                        "type": "string",
                        "description": "關卡 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "slot 索引",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.learnSlotReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/service.HintOutcome"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "410": {
                        "description": "Gone",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/learn/levels/{id}/reveal": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Learn"
                ],
                "summary": "揭曉答案",
                "parameters": [
                    {
                        "type": "string",
                        "description": "關卡 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "slot 索引",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.learnSlotReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/service.GuessOutcome"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "410": {
                        "description": "Gone",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/learn/srs": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Learn"
                ],
                "summary": "開始 SRS 複習 session",
                "parameters": [
                    {
                        "description": "題數與語系",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.startSRSReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/service.SRSSessionView"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/learn/srs/{id}/answer": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Learn"
                ],
                "summary": "作答 SRS 複習卡",
                "parameters": [
                    {
                        "type": "string",
                        "description": "session ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "卡片索引與作答",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.srsAnswerReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/service.SRSAnswerOutcome"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "410": {
                        "description": "Gone",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handler.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "handler.createCrosswordReq": {
            "type": "object",
            "required": [
                "tier"
            ],
            "properties": {
                "count": {
                    "description": "答案數（少3/中6/多9）；0 = 預設",
                    "type": "integer"
                },
                "locale": {
                    "type": "string"
                },
                "tier": {
                    "type": "integer"
                }
            }
        },
        "handler.createLevelReq": {
            "type": "object",
            "required": [
                "mode",
                "tier"
            ],
            "properties": {
                "count": {
                    "description": "答案數（少3/中6/多9）；0 = 該模式預設",
                    "type": "integer"
                },
                "locale": {
                    "description": "前端 i18n locale；空值 fallback en",
                    "type": "string"
                },
                "mode": {
                    "type": "string"
                },
                "tier": {
                    "type": "integer"
                }
            }
        },
        "handler.learnSlotReq": {
            "type": "object",
            "properties": {
                "slot": {
                    "type": "integer"
                }
            }
        },
        "handler.srsAnswerReq": {
            "type": "object",
            "properties": {
                "guess": {
                    "type": "string"
                },
                "index": {
                    "type": "integer"
                }
            }
        },
        "handler.startCampaignReq": {
            "type": "object",
            "properties": {
                "locale": {
                    "type": "string"
                }
            }
        },
        "handler.startSRSReq": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "locale": {
                    "type": "string"
                }
            }
        },
...
        "service.CrosswordView": {
            "type": "object",
            "properties": {
                "campaign": {
                    "description": "\u003e0 = 固定關卡編號",
                    "type": "integer"
                },
                "cols": {
                    "type": "integer"
                },
                "letters": {
                    "type": "string"
                },
                "level_id": {
                    "type": "string"
                },
                "rows": {
                    "type": "integer"
                },
                "tier": {
                    "type": "integer"
                },
                "words": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/service.CrosswordSlot"
                    }
                }
            }
        },
...
        "service.GuessOutcome": {
            "type": "object",
            "properties": {
                "completed": {
                    "type": "boolean"
                },
                "correct": {
                    "type": "boolean"
                },
                "definition": {
                    "type": "string"
                },
                "phonetic": {
                    "type": "string"
                },
                "slot": {
                    "type": "integer"
                },
                "total_xp": {
                    "type": "integer"
                },
                "word": {
                    "type": "string"
                },
                "xp_awarded": {
                    "type": "integer"
                }
            }
        },
...
        "service.HintOutcome": {
            "type": "object",
            "properties": {
                "definition": {
                    "type": "string"
                },
                "masked": {
                    "type": "string"
                },
                "slot": {
                    "type": "integer"
                },
                "tier": {
                    "type": "integer"
                }
            }
        },
...
        "service.LeaderboardView": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string"
                },
                "me": {
                    "description": "未上榜（未完成/未得分）時 nil",
                    "allOf": [
                        {
                            "$ref": "#/definitions/service.LeaderboardEntry"
                        }
                    ]
                },
                "top": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/service.LeaderboardEntry"
                    }
                },
                "week": {
                    "type": "string"
                }
            }
        },
        "service.LearnGuessRequest": {
            "type": "object",
            "required": [
                "word"
            ],
            "properties": {
                "slot": {
                    "type": "integer"
                },
                "word": {
                    "type": "string"
                }
            }
        },
...
        "service.LevelView": {
            "type": "object",
            "properties": {
                "letters": {
                    "type": "string"
                },
                "level_id": {
                    "type": "string"
                },
                "mode": {
                    "type": "string"
                },
                "slots": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/service.SlotView"
                    }
                },
                "tier": {
                    "type": "integer"
                }
            }
        },
...
        "service.SRSAnswerOutcome": {
            "type": "object",
            "properties": {
                "answer": {
                    "type": "string"
                },
                "completed": {
                    "type": "boolean"
                },
                "correct": {
                    "type": "boolean"
                },
                "next_stage": {
                    "type": "integer"
                },
                "phonetic": {
                    "type": "string"
                },
                "total_xp": {
                    "type": "integer"
                },
                "trans": {
                    "type": "string"
                },
                "xp_awarded": {
                    "type": "integer"
                }
            }
        },
...
        "service.SRSSessionView": {
            "type": "object",
            "properties": {
                "cards": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/service.SRSCardView"
                    }
                },
                "new_count": {
                    "type": "integer"
                },
                "session_id": {
                    "type": "string"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
    }
}